+++ src/app/pages/home/components/TodoHeader.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 
 import { TaskProps } from '../../../shared/models/task';
 import { Status } from '../../../shared/constants';
-import { addTask } from '../../../../redux/action';
+import { addTask } from '../../../redux/action';
 
 export const TodoHeader = () => {
   const inputRef = useRef<HTMLInputElement>(null);
