+++ src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import store from './redux/store';
+import store from './app/redux/store';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
